// Neptune.Web/src/app/shared/generated/model/custom-attribute-type-upsert-dto.ts
export class CustomAttributeTypeUpsertDto { 
    CustomAttributeTypeName: string;
    CustomAttributeDataTypeID: number;
    MeasurementUnitTypeID?: number | null;
    IsRequired: boolean;
    CustomAttributeTypePurposeID: number;
    CustomAttributeTypeDescription?: string | null;
    CustomAttributeTypeOptionsSchema?: string | null;
    CustomAttributeTypeDefaultValue?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}